pyproject.toml
[project]
name = "research-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.7.1",
    "dotenv>=0.9.9",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.20",
    "pytest>=9.0.3",
]
